refatorando metodo de fazer request
